Reverted moving model similarity.
